Merge Release/2024 q3 into master

+++ services/api/src/admin_org.py
@@ -9,7 +9,7 @@ def get_all_orgs():
     query = (
         "SELECT to_jsonb(row) "
         "FROM ("
-        "SELECT org.name, "
+        "SELECT org.name, org.email, "
         "(SELECT COUNT(*) FROM public.user_organization uo WHERE uo.organization_id = org.organization_id) num_users, "
         "(SELECT COUNT(*) FROM public.rsu_organization ro WHERE ro.organization_id = org.organization_id) num_rsus "
         "FROM public.organizations org"
@@ -22,6 +22,7 @@ def get_all_orgs():
         row = dict(row[0])
         org_obj = {}
         org_obj["name"] = row["name"]
+        org_obj["email"] = row["email"]
         org_obj["user_count"] = row["num_users"]
         org_obj["rsu_count"] = row["num_rsus"]
         return_obj.append(org_obj)
@@ -106,12 +107,17 @@ def get_modify_org_data(org_name):
 
 
 def check_safe_input(org_spec):
-    special_characters = "!\"#$%&'()*@-+,/:;<=>?[\]^`{|}~"
+    special_characters = "!\"#$%&'()*@-+,/:;<=>?[\\]^`{|}~"
     # Check all string based fields for special characters
     if any(c in special_characters for c in org_spec["orig_name"]):
         return False
     if any(c in special_characters for c in org_spec["name"]):
         return False
+    if org_spec["email"]:
+        if org_spec["email"] != "" and not admin_new_user.check_email(
+            org_spec["email"]
+        ):
+            return False
     for user in org_spec["users_to_add"]:
         if not admin_new_user.check_email(user["email"]):
             return False
@@ -134,14 +140,15 @@ def modify_org(org_spec):
     # Check for special characters for potential SQL injection
     if not check_safe_input(org_spec):
         return {
-            "message": "No special characters are allowed: !\"#$%&'()*+,./:;<=>?@[\]^`{|}~. No sequences of '-' characters are allowed"
+            "message": "No special characters are allowed: !\"#$%&'()*+,./:;<=>?@[\\]^`{|}~. No sequences of '-' characters are allowed"
         }, 500
 
     try:
         # Modify the existing organization data
         query = (
             "UPDATE public.organizations SET "
-            f"name = '{org_spec['name']}' "
+            f"name = '{org_spec['name']}', "
+            f"email = '{org_spec['email']}' "
             f"WHERE name = '{org_spec['orig_name']}'"
         )
         pgquery.write_db(query)
@@ -217,6 +224,17 @@ def modify_org(org_spec):
 
 
 def delete_org(org_name):
+
+    if check_orphan_rsus(org_name):
+        return {
+            "message": "Cannot delete organization that has one or more RSUs only associated with this organization"
+        }, 400
+
+    if check_orphan_users(org_name):
+        return {
+            "message": "Cannot delete organization that has one or more users only associated with this organization"
+        }, 400
+
     # Delete user-to-organization relationships
     user_org_remove_query = (
         "DELETE FROM public.user_organization WHERE "
@@ -235,7 +253,35 @@ def delete_org(org_name):
     org_remove_query = "DELETE FROM public.organizations WHERE " f"name = '{org_name}'"
     pgquery.write_db(org_remove_query)
 
-    return {"message": "Organization successfully deleted"}
+    return {"message": "Organization successfully deleted"}, 200
+
+
+def check_orphan_rsus(org):
+    rsu_query = (
+        "SELECT to_jsonb(row) "
+        "FROM (SELECT rsu_id, count(organization_id) count FROM rsu_organization WHERE rsu_id IN (SELECT rsu_id FROM rsu_organization WHERE organization_id = "
+        f"(SELECT organization_id FROM organizations WHERE name = '{org}')) GROUP BY rsu_id) as row"
+    )
+    rsu_count = pgquery.query_db(rsu_query)
+    for row in rsu_count:
+        rsu = dict(row[0])
+        if rsu["count"] == 1:
+            return True
+    return False
+
+
+def check_orphan_users(org):
+    user_query = (
+        "SELECT to_jsonb(row) "
+        "FROM (SELECT user_id, count(organization_id) count FROM user_organization WHERE user_id IN (SELECT user_id FROM user_organization WHERE organization_id = "
+        f"(SELECT organization_id FROM organizations WHERE name = '{org}')) GROUP BY user_id) as row"
+    )
+    user_count = pgquery.query_db(user_query)
+    for row in user_count:
+        user = dict(row[0])
+        if user["count"] == 1:
+            return True
+    return False
 
 
 # REST endpoint resource class
@@ -257,6 +303,7 @@ class UserRoleSchema(Schema):
 class AdminOrgPatchSchema(Schema):
     orig_name = fields.Str(required=True)
     name = fields.Str(required=True)
+    email = fields.Str(required=True, allow_none=True)
     users_to_add = fields.List(fields.Nested(UserRoleSchema), required=True)
     users_to_modify = fields.List(fields.Nested(UserRoleSchema), required=True)
     users_to_remove = fields.List(fields.Nested(UserRoleSchema), required=True)
@@ -313,4 +360,5 @@ class AdminOrg(Resource):
             abort(400, errors)
 
         org_name = urllib.request.unquote(request.args["org_name"])
-        return (delete_org(org_name), 200, self.headers)
+        message, code = delete_org(org_name)
+        return (message, code, self.headers)
